+++ tests/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path, include
 
 urlpatterns = [
-    path("billing/api/", include("stapel_billing.urls")),
+    path("billing/api/", include("stapel_billing.urls_v1")),
 ]